+++ tdishr/TdiConstant.c
@@ -15,7 +15,7 @@ STATIC_CONSTANT char *cvsrev = "@(#)$RCSfile$ $Revision$ $Date$";
 
 TdiRefStandard(Tdi1Constant)
 
-	status = MdsCopyDxXd((struct descriptor *)(*TdiRefFunction[opcode].f3)(), out_ptr);
+	status = MdsCopyDxXd((struct descriptor *)(long)(*TdiRefFunction[opcode].f3)(), out_ptr);
 	return status;
 }
 /*------------------------------------------------
